+++ framework/src/org/checkerframework/common/basetype/messages.properties
@@ -18,7 +18,6 @@ bound.type.incompatible=incompatible bounds in %s\ntype: %s\nupper bound: %s\nlo
 monotonic.type.incompatible=cannot assign %s to %s (monotonic type).\nfull type found: %s
 type.invalid=invalid type: annotations %s in type "%s"
 type.invalid.conflicting.annos=invalid type: conflicting annotations %s in type "%s"
-type.invalid.too.many.annotations=invalid type: too many annotations %s in type "%s"
 type.invalid.too.few.annotations=invalid type: missing annotations %s in type "%s"
 type.invalid.annotations.on.use=invalid type: annotations %s conflict with declaration of type %s
 cast.unsafe="%s" may not be casted to the type "%s"
