+++ framework/src/org/checkerframework/framework/util/AnnotatedTypes.java
@@ -39,8 +39,6 @@ import javax.lang.model.type.TypeVariable;
 import javax.lang.model.util.ElementFilter;
 import javax.lang.model.util.Elements;
 import javax.lang.model.util.Types;
-import org.checkerframework.framework.qual.PolyAll;
-import org.checkerframework.framework.source.Result;
 import org.checkerframework.framework.type.AnnotatedTypeFactory;
 import org.checkerframework.framework.type.AnnotatedTypeMirror;
 import org.checkerframework.framework.type.AnnotatedTypeMirror.AnnotatedArrayType;
@@ -52,7 +50,6 @@ import org.checkerframework.framework.type.AnnotatedTypeMirror.AnnotatedWildcard
 import org.checkerframework.framework.type.AsSuperVisitor;
 import org.checkerframework.framework.type.QualifierHierarchy;
 import org.checkerframework.framework.type.SyntheticArrays;
-import org.checkerframework.framework.type.visitor.SimpleAnnotatedTypeScanner;
 import org.checkerframework.javacutil.AnnotationUtils;
 import org.checkerframework.javacutil.ElementUtils;
 import org.checkerframework.javacutil.ErrorReporter;
@@ -868,78 +865,6 @@ public class AnnotatedTypes {
         }
 
         return hasTypeUse;
-    }
-
-    /**
-     * Returns true if the given {@link AnnotatedTypeMirror} passed a set of well-formedness checks.
-     * The method will never return false for valid types, but might not catch all invalid types.
-     *
-     * <p>Currently, the following is checked:
-     *
-     * <ol>
-     *   <li>There should not be multiple annotations from the same hierarchy.
-     *   <li>There should not be more annotations than the width of the qualifier hierarchy.
-     *   <li>If the type is not a type variable, then the number of annotations should be the same
-     *       as the width of the qualifier hierarchy.
-     *   <li>These properties should also hold recursively for component types of arrays, as wells
-     *       as bounds of type variables and wildcards.
-     * </ol>
-     */
-    public static Result isValidType(
-            QualifierHierarchy qualifierHierarchy, AnnotatedTypeMirror type) {
-        SimpleAnnotatedTypeScanner<Result, Void> scanner =
-                new SimpleAnnotatedTypeScanner<Result, Void>() {
-                    @Override
-                    protected Result defaultAction(AnnotatedTypeMirror type, Void aVoid) {
-                        return isTopLevelValidType(qualifierHierarchy, type);
-                    }
-
-                    @Override
-                    protected Result reduce(Result r1, Result r2) {
-                        if (r1 == null) {
-                            if (r2 == null) {
-                                return Result.SUCCESS;
-                            }
-                            return r2;
-                        } else if (r2 == null) {
-                            return r1;
-                        }
-                        return r1.merge(r2);
-                    }
-                };
-        return scanner.visit(type);
-    }
-
-    /** Checks every property listed in #isValidType, but only for the top level type. */
-    private static Result isTopLevelValidType(
-            QualifierHierarchy qualifierHierarchy, AnnotatedTypeMirror type) {
-        // multiple annotations from the same hierarchy
-        Set<AnnotationMirror> annotations = type.getAnnotations();
-        Set<AnnotationMirror> seenTops = AnnotationUtils.createAnnotationSet();
-        int n = 0;
-        for (AnnotationMirror anno : annotations) {
-            if (QualifierPolymorphism.isPolyAll(anno)) {
-                // ignore PolyAll when counting annotations
-                continue;
-            }
-            n++;
-            AnnotationMirror top = qualifierHierarchy.getTopAnnotation(anno);
-            if (AnnotationUtils.containsSame(seenTops, top)) {
-                return Result.failure("type.invalid.conflicting.annos", annotations, type);
-            }
-            seenTops.add(top);
-        }
-
-        // treat types that have polyall like type variables
-        boolean hasPolyAll = type.hasAnnotation(PolyAll.class);
-        boolean canHaveEmptyAnnotationSet =
-                QualifierHierarchy.canHaveEmptyAnnotationSet(type) || hasPolyAll;
-
-        // wrong number of annotations
-        if (!canHaveEmptyAnnotationSet && seenTops.size() < qualifierHierarchy.getWidth()) {
-            return Result.failure("type.invalid.too.few.annotations", annotations, type);
-        }
-        return Result.SUCCESS;
     }
 
     private static String annotationClassName =
